some refactoring
put everything to do with ordering largest chats in the largest chat file
